+++ SurfaceTopography/ChangeLog.md
@@ -1,9 +1,11 @@
 Change log for SurfaceTopography
 =============================
 
-v1.2.3 (not yet released)
--------------------------
+v1.2.3 (01Dec22)
+----------------
 
+- BUG: Make sure all readers return the same info dictionary for channels
+  and the final topography
 - MAINT: Automatic version discovery for Meson build
 
 v1.2.2 (28Nov22)